refactor: extract shared utilities and fix README duplication
- Create cli-automation-utils.js: Shared utilities for terminal session, command execution, and output processing
- Update create-cli-videos.js: Use shared utilities, remove duplicated functions
- Update create-cli-screenshots.js: Use shared utilities, remove duplicated functions
- Fix README.md: Consolidate Quick Start and CLI Usage sections to eliminate duplication
- Improve code maintainability: Single source of truth for automation functionality
- Enhance consistency: Standardized terminal styling and command execution across scripts

Addresses PR review comments requesting shared utility extraction and documentation consolidation.
Reduces code duplication by ~300 lines while maintaining full functionality.
